requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyMuPDF==1.24.9
python-docx==1.1.2